+++ docs/examples/parallax/lib/main.dart
@@ -4,6 +4,7 @@ import 'package:flame/components/parallax_component.dart';
 import 'package:flutter/material.dart';
 
 void main() async {
+  WidgetsFlutterBinding.ensureInitialized();
   await Flame.util.fullScreen();
   runApp(MyGame().widget);
 }