+++ scripts/lint.sh
@@ -5,6 +5,7 @@ fi
 
 result=`dartanalyzer lib/`
 if ! echo "$result" | grep -q "No issues found!"; then
+  echo "$result"
   echo "dartanalyzer issue"
   exit 1
 fi