+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [next]
 - Updated the doc structure and minor language fixes
 - Adding AssetsCache.readBinaryFile
+- Splitting debugMode from recordFps mode
 
 ## 0.20.2
 - Fix text component bug with anchor being applied twice