+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 
 ## [next]
+- Updated the doc structure and minor language fixes
+- Adding AssetsCache.readBinaryFile
 
 ## 0.20.2
 - Fix text component bug with anchor being applied twice