+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [next]
 - Updated the doc structure and minor language fixes
 - Adding AssetsCache.readBinaryFile
+- Splitting debugMode from recordFps mode
 - Adding support for multi touch tap and drag events
 - Fix animations example
 - Add possibility for infinite and alternating effects